mobile concluido faltando relatorios

+++ Backend/src/main/java/com/mensal/compras/dto/ItemWithdrawalDTO.java
@@ -1,6 +1,7 @@
 package com.mensal.compras.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,7 +15,7 @@ public class ItemWithdrawalDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	private Long id;	
-	private Double quantity;
+	private BigDecimal quantity;
 	private Long productId;
 	
 }
